Updated folder structure and renamed the following
LoginViewModel -> AuthenticationViewModel

ChangePassword -> ForgetPassword
